Move reader_base to its own .h and .cpp
